+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Update dependencies (jest-cli@0.9.0)
 - Improve lazy dependency installation
     - Allow global installation
         use ```global``` parameter of ```install-dependencies``` command