+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.6.0 (2016-03-28)
 - Introduce ```check-dependencies``` task
 - Improve lazy dependency installation
     - Allow global installation